Feature/fixup tests for easier rebasing (#308)
* Rename files in order to simplify rebase

* Move initialization into beforeEach

* Update default sub claim in testAuthData

* Fix tyos as suggested


---------
