+++ __tests__/api/league/route.js
@@ -2,13 +2,46 @@
  * @jest-environment node
  */
 
+jest.mock("google-auth-library");
+import { OAuth2Client } from "google-auth-library";
 import { POST } from "@/app/api/league/route.ts";
 
-describe("POST", () => {
-    it("should return a 401 when auth token is missing", async () => {
-        // Aarrange
+let testAuthData = {}
+
+const getPayloadMock = jest.fn().mockImplementation(()=> {
+    return testAuthData
+});
+
+const verifyIdTokenMock = jest.fn().mockImplementation(()=> {
+    return {
+        getPayload: getPayloadMock
+    }
+});
+
+OAuth2Client.mockImplementation(() => {
+    return {
+        verifyIdToken: verifyIdTokenMock
+    }
+});
+
+beforeEach(() => {
+    testAuthData = {
+        sub: "108251633753098119380",
+        email: "test@test.com",
+        given_name: "TestFirstName",
+        family_name: "TestLastName"
+    }
+});
+
+describe("POST (unit tests)", () => {
+
+    it("should return a 403 when auth token does not have exact right userId claim", async () => {
+        // Arrange
+        testAuthData.sub = "123googleTestId";
         const request = {
-            json: async () => { return {} }
+            json: async () => { return {
+                token: "testToken"
+            } }
         };
 
         // Act
@@ -16,13 +49,15 @@ describe("POST", () => {
 
         // Assert
         expect(response).not.toBeNull();
-        expect(response.status).toEqual(401);
+        expect(response.status).toEqual(403);
     });
 
-    it("should return a 401 when auth token is malformed (missing 3 parts)", async () => {
+    it("should return allow a post (or  200) when auth token has the exact right userId claim", async () => {
         // Arrange
         const request = {
-            json: async () => { return { token: "someTokenValue" } }
+            json: async () => { return {
+                token: "testToken"
+            } }
         };
 
         // Act
@@ -30,20 +65,6 @@ describe("POST", () => {
 
         // Assert
         expect(response).not.toBeNull();
-        expect(response.status).toEqual(401);
-    });
-
-    it("should return a 401 when auth token is malformed (parts are not base64 encoded json)", async () => {
-        // Aarrange
-        const request = {
-            json: async () => { return { token: "someTokenHeader.someTokenBody.someTokenSig" } }
-        };
-
-        // Act
-        const response = await POST(request);
-
-        // Assert
-        expect(response).not.toBeNull();
-        expect(response.status).toEqual(401);
+        expect(response.status).toEqual(200);
     });
 });
